v102
If you properly exit the program the items on your to do list will be saved to a pre-defined .txt file.
When you run the program again you can list the saved items.
When completing a task you get a confirmation message.
If you already completed the task and you try to complete it again you are told that you already completed that task.