Multi-user support completed
V1.4 - Multi-user support
Multiple users want to use the program. Users are identified by their username (string). ✓
When a task is added, the username needs to be specified. ✓
The program needs to be able to list all tasks for a given user. ✓
When tasks are listed without specifying a username all tasks should be listed grouped by the username. ✓
The remove command should also ask for a username in order to know which user is removing the task.✓