Task 1 - 3
I managed to get the ID bits work to the extent where all new tasks get an ID and the user can see the ID when lists the tasks on the to do list. Some issues persist, e.g. if you delete a task  I need to figure out how to show the appropriate task ID to the user.